Fix a bug related to deleting image when it is used by a container
Before, when an image is referenced by a container, user is still
able to delete the image even without force command

To avoid that thing to happen, another column is added
in the delete message(shown below):

 IMAGE                     STORAGE       USED
 xxxxx                     xxxxxxx       xxxx

When user is trying to delete an image,
the USED column will have a YES or NO answer
to indicate the stauts of the image

The user will no longer be able to delete
"used" images(without force)
and an error message will be
shown to further indicate that.

Some tests are added for future regression.