Add dangling property checking and Fix a bug to prevent deleting not dangling unnamed layers
Before, the function for dangling property for system containers
were empty, now the function was added and is_dangling property
can be checked via the *(which was originally there)

Before, users were able to delete unnamed layer of an image
when the named layer image is still present. With the help
of is_dangling property, if the repo for the image is <none>
(which indicates that it is an unnamed layer of image), a
further check was applied to see if the image is dangling or not.
If it is not dangling, it is prevented from the user to deleting the
image

Before, when the user types in a abbreviation of image id
of an unnamed layer image, the image id shown on the command
line interface was off(the commit id was shown instead).

Some integration tests are added for future regression