syscontainer: remove of fallback version of skopeo
Based on discussion, skopeo is assumed to include support for
ostree copy from now on.

The users will now be noted to upgrade skopeo to newest version
if their skopeo version does not support copy to ostree.

For now, looks like user will be able to use skopeo copy with ostree
after version 0.1.20, so skopeo v0.1.21 should contain the feature
https://github.com/projectatomic/skopeo/commit/1d5c681f0fc6260d7585b65345342ab64c771b17