syscontainer: fix for missing images display in ostree backend
This fixes https://github.com/projectatomic/atomic/issues/1179.

Before, we assumed the image ref with 64 characters is default to
image layers(images with repository <none>), but as issue 1179
pointed out, there are some exceptions.

We now migrated to check if the ref is hex. If the ref is hex, we
can conclude it should be images with repository <none>