syscontainer: catch missing container info failure
As reported by https://github.com/projectatomic/atomic/issues/1191,
when info file is missing, atomic images list will unexpectedly fail.

This patch is therefore to catch the error and instead output a
warning to notify user.

This patch also adds an error handling case for following commands:
1: atomic uninstall $container
2: atomic containers update $container
3: atomic containers rollback $container

Now instead of erroring out, a warning is added to notify the user
about the failure